lru-dict-memleak-test/requirements.txt
lru-dict==1.1.6
psutil
